:arrow_up:  嘗試更新RN版本0.71.0
