+++ Gemfile
@@ -1,7 +1,6 @@
 source 'https://rubygems.org'
 
 # You may use http://rbenv.org/ or https://rvm.io/ to install and use this version
-ruby '>= 2.6.10'
+ruby File.read(File.join(__dir__, '.ruby-version')).strip
 
-gem 'cocoapods', '>= 1.11.3'
-gem 'activesupport', '>= 6.1.7.3', '< 7.1.0'
+gem 'cocoapods', '~> 1.11', '>= 1.11.3'
