+++ android/settings.gradle
@@ -3,11 +3,6 @@ apply from: file("../node_modules/@react-native-community/cli-platform-android/n
 include ':app'
 includeBuild('../node_modules/react-native-gradle-plugin')
 
-if (settings.hasProperty("newArchEnabled") && settings.newArchEnabled == "true") {
-    include(":ReactAndroid")
-    project(":ReactAndroid").projectDir = file('../node_modules/react-native/ReactAndroid')
-}
-
 // 漸變庫安裝需要
 include ':react-native-linear-gradient'
 project(':react-native-linear-gradient').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-linear-gradient/android')
@@ -32,8 +27,8 @@ include ':react-native-compressor'
 project(':react-native-compressor').projectDir = new File(rootProject.projectDir,'../node_modules/react-native-compressor/android')
 
 // 更新gradle時，https://reactnative.dev/docs/new-architecture-app-intro
-includeBuild('../node_modules/react-native-gradle-plugin')
-include(":ReactAndroid")
-project(":ReactAndroid").projectDir = file('../node_modules/react-native/ReactAndroid')
-include(":ReactAndroid:hermes-engine")
-project(":ReactAndroid:hermes-engine").projectDir = file('../node_modules/react-native/ReactAndroid/hermes-engine')
+// includeBuild('../node_modules/react-native-gradle-plugin')
+// include(":ReactAndroid")
+// project(":ReactAndroid").projectDir = file('../node_modules/react-native/ReactAndroid')
+// include(":ReactAndroid:hermes-engine")
+// project(":ReactAndroid:hermes-engine").projectDir = file('../node_modules/react-native/ReactAndroid/hermes-engine')
