+++ package.json
@@ -7,9 +7,9 @@
     "ios": "react-native run-ios --simulator=\"iPhone 15 Pro Max\"",
     "iosTrue": "react-native run-ios --device",
     "iosBig": "react-native run-ios --simulator=\"iPad Pro 13-inch (M4)\"",
+    "lint": "eslint .",
     "start": "react-native start --reset-cache",
     "test": "jest",
-    "lint": "eslint .",
     "postinstall": "patch-package"
   },
   "dependencies": {
@@ -41,10 +41,10 @@
     "moment-timezone": "^0.5.45",
     "patch-package": "^8.0.0",
     "qs": "^6.13.0",
-    "react": "18.1.0",
+    "react": "18.2.0",
     "react-content-loader": "^6.2.1",
     "react-i18next": "^14.1.3",
-    "react-native": "0.70.15",
+    "react-native": "0.71.0",
     "react-native-animated-nav-tab-bar": "^3.1.11",
     "react-native-animated-splash-screen": "^2.0.5",
     "react-native-blob-util": "^0.19.11",
@@ -91,13 +91,20 @@
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.14.0",
     "@babel/runtime": "^7.25.6",
     "@react-native-community/eslint-config": "^3.2.0",
+    "@tsconfig/react-native": "^2.0.2",
+    "@types/jest": "^29.2.1",
+    "@types/react": "^18.0.24",
+    "@types/react-test-renderer": "^18.0.0",
     "babel-jest": "^29.7.0",
     "eslint": "^8.57.1",
     "jest": "^29.7.0",
     "metro-react-native-babel-preset": "^0.77.0",
-    "react-test-renderer": "18.1.0"
+    "prettier": "^2.4.1",
+    "react-test-renderer": "18.2.0",
+    "typescript": "4.8.4"
   },
   "jest": {
     "preset": "react-native"
